backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for Lokalize AI Translator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.3",
    "pydantic-settings>=2.2.1",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "python-lokalise-api>=2.0.0",
    "google-generativeai>=0.8.0",
    "tenacity>=8.2.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "python-multipart>=0.0.9",
    "sacrebleu>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"

# Exclude a variety of commonly ignored directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
# Enable all rules by default
select = ["E", "F", "B", "I", "N", "UP", "PL", "RUF"]

# Exclude specific rules
ignore = [
    "E501",  # Line too long (handled by black)
    "PLR0913",  # Too many arguments to function call
]

# Allow autofix for all enabled rules
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.flake8-unused-arguments]
ignore-variadic-names = true

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "validator"]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
strict = true

[tool.basedpyright]
venvPath = "."
venv = ".venv"
typeCheckingMode = "basic"
useLibraryCodeForTypes = true
pythonPath = ".venv/bin/python"
extraPaths = [".venv/lib/python3.13/site-packages"]

[dependency-groups]
dev = [
    "basedpyright>=1.29.4",
]

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["logs*", "tests*"]

[tool.uv.workspace]
members = [
    "lokalize-ai-translator",
]
